♻️ Refactor: my-profile 파일 분리
커스텀 훅을 만들어서 함수들 정리
